Improve extract script

+++ SpatialSense/extractCode.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
-rm -rf code
-tar -xvf code.tar
-rm code.tar
+test -e code.tar
+
+if [[ $? -ne 0 ]]; then
+	echo "no tar file, aborting"
+else
+	rm -rf code
+	tar -xvf code.tar
+	rm code.tar
+fi
